Move `bspc` and `tab` to right-thumb positions
BSPC is used relatively frequently, and when used is often used in rapid
succession. Moving this action away from the pinky and onto the thumb
will balance the stress better.

Tab is often used for command completion at the command line, some of
which is path completion. Since path completion uses a lot of slash
characters, which is currently mapped to the pinky, having tab be on the
pinky as well really slows things down and requires a lot of movement.
Moving it to a thumb button alleviates that.

Furthermore this frees up pinky spaces for modifiers, allowing LAlt and
LGUI to be moved away from the thumb, making it easier to hit modifier
combos.