Remove duplicated vim-yank-to-clipboard (FN4) key
The yank to clipboard FN is FN4, which used to be mapped to a key next
to backspace despite not being used. This results in erroneously hitting
the key due to fat-finger errors, which is annoying and can easily be
fixed by making it blank (or, in this case, a blank FN, ready to be
mapped at a later date).